feat: Provide active CSS class to current page
This replaces the old method of checking based on the page name, which
generally increased the size of all pages (and the CSS) by a couple
dozen bytes. Nav pages are also determined programmatically from the
data file.

The old CSS associated with the nav has been marked as depreciated.

Autoformatter hit the default layout, sorry about that.